fix(migration): fix migration for dev env

+++ packages/backend/src/_migrations/1671120114644-auto-migration.ts
@@ -1,10 +1,15 @@
 import { MigrationInterface, QueryRunner } from "typeorm";
+import { domifaConfig } from "../config";
 
 export class autoMigration1671120114644 implements MigrationInterface {
   name = "autoMigration1671120114644";
 
   public async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query(`ALTER TABLE "usager" DROP COLUMN "oldNotes"`);
+    if (domifaConfig().envId !== "test" && domifaConfig().envId !== "dev") {
+      await queryRunner.query(`ALTER TABLE "usager" DROP COLUMN "oldNotes"`);
+    } else {
+      console.log("Migration done ");
+    }
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
